fix(deps): update cargo and npm dependencies, closing 6 advisories
Brings both ecosystems to latest; `npm outdated` is now empty and both audits
report zero vulnerabilities (npm was at 6 — three of them high).

The npm advisories were `postcss` path traversal, `fast-uri` host confusion,
`esbuild` dev-server file read, and fourteen `dompurify` XSS reports. dompurify
had no clean route: the newest monaco still pins a vulnerable version, and npm's
own suggested fix was a *downgrade* to monaco 0.53. It is pinned to 3.4.12
through an override instead, so monaco could move forward rather than back.

Notable constraints and knock-on effects:

- rusqlite stops at 0.39, not 0.40 — `refinery-core` requires `<=0.39`, so the
  migration runner sets the ceiling. rusqlite also dropped `ToSql for usize`
  (its width is platform-dependent), so the two LIMIT binds cast to i64.
- reqwest 0.13 renamed `rustls-tls` to `rustls`, which swaps the crypto provider
  from ring to aws-lc-rs and so adds a cmake/C build to the musl static runner.
  Verified against a CI-shaped Ubuntu-to-musl cross-build; `build.yml` now names
  cmake explicitly rather than relying on the runner image preinstalling it,
  because that job runs only on master and tags — the failure would surface on a
  release, not on a PR.
- keyring 4 restructures into keyring-core plus per-platform stores. The `v1`
  feature is the same three stores the explicit list named, and registers them
  lazily on first `Entry::new`. Its *nix backend is now pure-Rust zbus rather
  than libdbus, which retired the reason recorded next to the feature flag.
- monaco 0.56 added an `exports` map, making the deep `esm/vs/...` worker paths
  unresolvable. Vite 7 tolerated it; Vite 8's Rolldown does not, so the imports
  now go through the exports map.
- jsonschema turned `ValidationError::instance_path` into a method.
- The vendored quick-xml patch is deleted. It existed to dodge
  RUSTSEC-2026-0194/0195 while `plist` still required `^0.39.2`; plist 1.10.0
  pulls the fixed 0.41.0 from the registry, so the workaround was inert.
- `xterm` and `xterm-addon-fit` are removed — deprecated upstream and
  unreferenced since everything moved to `@xterm/*`.

TypeScript 7 and vitest 4 needed no source changes at all.

The new migration tests cover an upgrade path no fresh-database test can
exercise. refinery records a per-migration checksum and aborts by default when a
stored one disagrees, which across a version bump would mean an app that refuses
to open every database already on disk. `migration::run` disables that, and the
tests pin it by migrating for real and then rewriting the recorded checksums.

+++ crates/demeteo-core/src/adapters/database/repos/memory.rs
@@ -78,7 +78,7 @@ impl ProjectMemoryPort for SqliteAdapter {
         );
         let mut stmt = conn.prepare(&sql).map_err(|e| e.to_string())?;
         let iter = stmt
-            .query_map(params![project_id, limit], row_to_entry)
+            .query_map(params![project_id, limit as i64], row_to_entry)
             .map_err(|e| e.to_string())?;
 
         let mut list = Vec::new();
